requirements.txt
azure-communication-chat==1.1.0
azure-communication-identity==1.0.1
azure-core==1.19.0
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
Flask==2.0.2
idna==3.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
msrest==0.6.21
oauthlib==3.1.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
